+++ site/dat/docs/Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.4.6 (next)
+## 0.5.0 (next)
  - allow using Xvfb for Selenium executor
  - don't use lynx browser when openin report links
  - add PBench executor type